requirements.txt
SpeechRecognition==3.8.1
Flask==1.1.2
google-cloud-speech